third_party/blink/web_tests/html/selectmenu/selectmenu-typeahead-scroll-expected.txt

PASS selectmenu.open is true
PASS successfullyParsed is true

TEST COMPLETE


Test that selection box scrolls to where the focus jumps when pressing an alphanumeric key.

Button
aardvark
blue
navy
indigo
azure
SUCCESS
